requirements.txt
aiocfscrape==1.0.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfscrape==2.1.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
constantly==15.1.0
cryptography==41.0.1
cssselect==1.2.0
debugpy==1.6.7
decorator==5.1.1
environs==9.5.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.2
frozenlist==1.3.3
futures==3.0.5
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gspread==5.7.2
gspread-asyncio==1.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.23.2
ipython==8.14.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
jmespath==1.0.1
Js2Py==0.74
jupyter_client==8.2.0
jupyter_core==5.3.1
kaitaistruct==0.10
lxml==4.9.2
magic-filter==1.0.9
marshmallow==3.19.0
matplotlib-inline==0.1.6
multidict==6.0.4
mypy==1.4.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.8
outcome==1.2.0
packaging==23.1
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.7.0
prompt-toolkit==3.0.38
Protego==0.2.1
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pydivert==2.1.0
Pygments==2.15.1
pyjsparser==2.7.1
pyOpenSSL==23.2.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
Scrapy==2.9.0
selenium==4.10.0
selenium-wire==5.1.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
tldextract==3.4.4
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
undetected-chromedriver==3.5.0
urllib3==1.26.16
user-agent==0.1.10
w3lib==2.1.1
wcwidth==0.2.6
webdriver-manager==3.8.6
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.0
zstandard==0.21.0
